# docs/api/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QualiDiagIQ API",
    "description": "\n        Motor de Diagnóstico Tributário Automatizado para a Reforma do Consumo.\n        \n        Parte do ecossistema Tributiq. Fornece análise de maturidade e conformidade\n        com base na EC 132/2023, LC 214/2025 e ABNT NBR 17301:2026.\n        ",
    "contact": {
      "name": "Equipe Tributiq",
      "email": "contato@tributiq.com.br"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/diagnosticos/rascunho-self-service/vincular-conta": {
      "post": {
        "tags": [
          "Diagnósticos"
        ],
        "summary": "Materializar rascunho no tenant da conta (JWT)",
        "description": "Exige **Bearer** da conta na plataforma. O e-mail do respondente no rascunho deve ser **igual** ao e-mail do admin (LGPD / prova de posse). **Idempotency-Key** obrigatório.\n\n**CNPJ:** o payload do rascunho deve incluir CNPJ válido — ao vincular à conta o diagnóstico passa a exigir identificação PJ para histórico por empresa no tenant (ADR-013).",
        "operationId": "vincular_rascunho_conta_plataforma_diagnosticos_rascunho_self_service_vincular_conta_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VincularRascunhoContaPlataformaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/diagnosticos/self-service/conclusao-visualizacao": {
      "get": {
        "tags": [
          "Diagnósticos"
        ],
        "summary": "Visualização pública do diagnóstico concluído (self-service)",
        "description": "Endpoint **público** (sem JWT). Exige ``diagnostico_id`` e ``leitura_token`` devolvidos por **POST /diagnosticos/rascunho-self-service/concluir** (token persistido em PostgreSQL, TTL ~7 dias). Não expõe checklist/PDF — apenas snapshot executivo alinhado à página de conclusão.",
        "operationId": "obter_conclusao_self_service_publica_diagnosticos_self_service_conclusao_visualizacao_get",
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID do diagnóstico gravado no tenant self-service.",
              "title": "Diagnostico Id"
            },
            "description": "UUID do diagnóstico gravado no tenant self-service."
          },
          {
            "name": "leitura_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "description": "Token opaco devolvido no campo `leitura_token` da resposta de concluir.",
              "title": "Leitura Token"
            },
            "description": "Token opaco devolvido no campo `leitura_token` da resposta de concluir."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticoConclusaoSelfServicePublicoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/diagnosticos/vincular-leads-self-service": {
      "post": {
        "tags": [
          "Diagnósticos"
        ],
        "summary": "Vincular diagnósticos gratuitos (self-service) ao tenant da conta na plataforma",
        "description": "Reatribui ao tenant do JWT as linhas em `diagnosticos` gravadas no tenant self-service (fluxo OTP), com `respondente_email` igual ao e-mail do consultor em `admins` e plano gratuito. Resolve o caso em que o lead concluiu o assistente antes de iniciar sessão no painel. **Idempotency-Key** obrigatório (mesmo middleware dos outros POST sob `/diagnosticos/`).",
        "operationId": "vincular_leads_self_service_diagnosticos_vincular_leads_self_service_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VincularLeadsSelfServiceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/diagnosticos/": {
      "get": {
        "tags": [
          "Diagnósticos"
        ],
        "summary": "Listar Diagnosticos",
        "description": "Lista diagnósticos do tenant atual (ordenacao: mais recentes primeiro na camada repo/DB).",
        "operationId": "listar_diagnosticos_diagnosticos__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "empresa_cnpj",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opcional: filtra pela coluna `empresa_cnpj` (14 dígitos, DV válido). Omite ou vazio = todos os diagnósticos do tenant.",
              "title": "Empresa Cnpj"
            },
            "description": "Opcional: filtra pela coluna `empresa_cnpj` (14 dígitos, DV válido). Omite ou vazio = todos os diagnósticos do tenant."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiagnosticoResumoSchema"
                  },
                  "title": "Response Listar Diagnosticos Diagnosticos  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Diagnósticos"
        ],
        "summary": "Criar diagnóstico",
        "description": "Calcula o score e persiste o diagnóstico no tenant do JWT.\n\n**Headers obrigatórios:** `Authorization: Bearer <JWT>` (claim `tenant_id`) e `Idempotency-Key` (UUID v4 recomendado). Reexecução com a mesma chave devolve a mesma resposta 2xx em cache (middleware de idempotência).\n\n**Corpo:** incluir `aceite_termos_privacidade: true` (LGPD); o servidor persiste `aceite_termos_privacidade_em` (UTC) na linha do diagnóstico.\n\n**CNPJ:** obrigatório com sessão na plataforma (histórico por empresa no tenant). Fluxo só com OTP sem conta: usar rotas self-service / rascunho onde o CNPJ pode ser opcional.",
        "operationId": "criar_diagnostico_diagnosticos__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IniciarDiagnosticoPainelRequest"
              },
              "examples": {
                "micro_comercio_um_item": {
                  "summary": "Micro comércio — uma resposta escala",
                  "description": "Headers obrigatorios: Authorization Bearer (JWT valido do login) e Idempotency-Key (UUID v4). Resposta HTTP 201 com score e consultoria derivada.",
                  "value": {
                    "empresa": {
                      "cnpj": "12345678000195",
                      "razao_social": "Empresa Exemplo LTDA",
                      "porte": "micro",
                      "regime": "simples_nacional",
                      "cnae_principal": "1234567",
                      "uf": "SP",
                      "setor_macro": "comercio"
                    },
                    "respondente": {
                      "email": "fiscal@empresa.com.br",
                      "nome": "Fulano da Silva",
                      "telefone": "11999998888"
                    },
                    "respostas": [
                      {
                        "pergunta_id": "df52b20e-dab8-5ce5-a89d-4fb235016cbe",
                        "valor": 4
                      }
                    ],
                    "plano": "gratuito",
                    "locale_relatorio": "pt-BR",
                    "aceite_termos_privacidade": true
                  }
                },
                "servicos_medio_presumido": {
                  "summary": "Serviços — médio porte, Lucro Presumido",
                  "description": "Mesmo requisito de cadastro PJ: CNPJ válido (14 dígitos) + razão social e perfil fiscal.",
                  "value": {
                    "empresa": {
                      "cnpj": "11222333000181",
                      "razao_social": "Empresa Anônima Exemplo LTDA",
                      "porte": "medio",
                      "regime": "lucro_presumido",
                      "cnae_principal": "6201501",
                      "uf": "SP",
                      "setor_macro": "servicos"
                    },
                    "respondente": {
                      "email": "contador@escritorio.com.br",
                      "nome": "Maria Souza"
                    },
                    "respostas": [
                      {
                        "pergunta_id": "df52b20e-dab8-5ce5-a89d-4fb235016cbe",
                        "valor": 3
                      }
                    ],
                    "plano": "gratuito",
                    "locale_relatorio": "pt-BR",
                    "aceite_termos_privacidade": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiagnosticoConclusaoSelfServicePublicoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "empresa_razao_social": {
            "type": "string",
            "title": "Empresa Razao Social"
          },
          "locale_relatorio": {
            "type": "string",
            "title": "Locale Relatorio",
            "default": "pt-BR"
          },
          "score_geral": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Geral"
          },
          "scores_por_dimensao": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticoConclusaoPublicaDimensaoSchema"
            },
            "type": "array",
            "title": "Scores Por Dimensao"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "empresa_razao_social"
        ],
        "title": "DiagnosticoConclusaoSelfServicePublicoResponse",
        "description": "Snapshot mínimo para a página pública de conclusão (sem JWT).\n\nDados lidos de ``diagnosticos`` após validação do token de leitura na tabela dedicada."
      },
...
      "DiagnosticoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "plano": {
            "type": "string",
            "title": "Plano"
          },
          "empresa_razao_social": {
            "type": "string",
            "title": "Empresa Razao Social"
          },
          "empresa_cnpj": {
            "type": "string",
            "title": "Empresa Cnpj",
            "description": "CNPJ no snapshot do diagnóstico (14 dígitos ou vazio se não informado).",
            "default": ""
          },
          "criado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Em",
            "description": "Instante de criação (UTC) — útil para linha do tempo e comparativo entre ciclos."
          },
          "finalizado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalizado Em",
            "description": "Instante de finalização (UTC), se aplicável."
          },
          "empresa_faixa_faturamento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empresa Faixa Faturamento",
            "description": "Faixa de faturamento autodeclarada no POST (slug canónico), se informada."
          },
          "locale_relatorio": {
            "type": "string",
            "title": "Locale Relatorio",
            "description": "Idioma usado na geração do PDF (persistido com o diagnóstico).",
            "default": "pt-BR"
          },
          "score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoreCompletoSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "relatorio_pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relatorio Pdf Url"
          },
          "recomendacao_ia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recomendacao Ia"
          },
          "checklist": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checklist"
          },
          "matriz_impacto": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matriz Impacto"
          },
          "cronograma": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cronograma"
          },
          "checklist_m12_autoconf": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checklist M12 Autoconf"
          },
          "quadro_implantacao_anotacoes": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quadro Implantacao Anotacoes",
            "description": "Mapa f{i}_a{j} -> {prazo_meta, comentarios[]} (legado comentario migrado na leitura)."
          },
          "aceite_termos_privacidade_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aceite Termos Privacidade Em"
          },
          "hash_evidencia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hash Evidencia"
          },
          "versao_otimista": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Versao Otimista"
          },
          "versao_plano": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Versao Plano",
            "description": "Versão do snapshot do plano de ação materializado (checklist/matriz/cronograma).",
            "default": 1
          },
          "leitura_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leitura Token",
            "description": "Presente só em POST /diagnosticos/rascunho-self-service/concluir: token opaco para GET /diagnosticos/self-service/conclusao-visualizacao (leitura pública limitada no tempo)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "plano",
          "empresa_razao_social"
        ],
        "title": "DiagnosticoResponse"
      },
      "DiagnosticoResumoSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "empresa_razao_social": {
            "type": "string",
            "title": "Empresa Razao Social"
          },
          "empresa_cnpj": {
            "type": "string",
            "title": "Empresa Cnpj",
            "description": "CNPJ normalizado (14 dígitos) ou vazio — agrupa ciclos «antes/depois» no painel.",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "plano": {
            "type": "string",
            "title": "Plano"
          },
          "score_geral": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Geral"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "finalizado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalizado Em"
          },
          "relatorio_pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relatorio Pdf Url",
            "description": "URL do PDF quando gerado e anexado ao diagnóstico."
          }
        },
        "type": "object",
        "required": [
          "id",
          "empresa_razao_social",
          "status",
          "plano",
          "criado_em"
        ],
        "title": "DiagnosticoResumoSchema",
        "description": "Item resumido para listagem do tenant (P7 — painel)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScoreCompletoSchema": {
        "properties": {
          "score_geral": {
            "$ref": "#/components/schemas/ScoreDimensaoSchema"
          },
          "score_por_dimensao": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ScoreDimensaoSchema"
            },
            "type": "object",
            "title": "Score Por Dimensao"
          }
        },
        "type": "object",
        "required": [
          "score_geral",
          "score_por_dimensao"
        ],
        "title": "ScoreCompletoSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VincularLeadsSelfServiceResponse": {
        "properties": {
          "total_vinculados": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Vinculados",
            "description": "Quantidade de diagnósticos reatribuídos ao tenant da conta na plataforma."
          },
          "diagnostico_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Diagnostico Ids",
            "description": "Identificadores atualizados (ordem não garantida)."
          }
        },
        "type": "object",
        "required": [
          "total_vinculados"
        ],
        "title": "VincularLeadsSelfServiceResponse",
        "description": "Resposta de POST /diagnosticos/vincular-leads-self-service (reatribuição de tenant)."
      },
      "VincularRascunhoContaPlataformaRequest": {
        "properties": {
          "resgate_token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 16,
            "title": "Resgate Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "resgate_token"
        ],
        "title": "VincularRascunhoContaPlataformaRequest",
        "description": "Associa rascunho self-service ao tenant do JWT (e-mail do respondente = e-mail do admin)."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}